urbem_interface/requirements.txt
numpy>=1.20
pandas>=1.3
xarray>=2022.1
rasterio>=1.2
geopandas>=0.10
pyarrow>=12.0
pyogrio>=0.7
osmnx>=1.2
shapely>=2.0
flask>=2.0
pywebview>=4.0
Pillow>=9.0
tqdm>=4.60
